+++ Cargo.toml
@@ -9,3 +9,4 @@ edition = "2021"
 arboretum-td = "*"
 fxhash = "*"
 clap = { version = "4.0", features = ["derive"] }
+bit-set = "*"
