+++ src/graph.rs
@@ -7,6 +7,10 @@ pub mod quick_graph;
 pub(crate) mod reducible;
 pub(crate) mod reductions;
 pub mod tree_decomposition;
+pub mod nice_tree_decomp;
+pub mod node_relations;
+pub mod mis_finder;
+pub mod dyn_table;
 
 pub use builder::dcel_builder::DcelBuilder;
 pub use dcel::Dcel;
