+++ src/graph/dyn_table.rs
@@ -1,0 +1,2 @@
+pub mod dt_fast;
+pub mod dt_normal;
